requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
python-multipart==0.0.20
tensorflow==2.20.0
tf-keras==2.20.1
numpy==1.26.4
opencv-python==4.8.1.78
mtcnn==1.0.0
pillow==12.0.0
deepface==0.0.96